refactor: create offcanvas folder and move createContract be a component
